% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_utils.R
\name{plot.vel_bmgm}
\alias{plot.vel_bmgm}
\title{Plot Method for VEL-BMGM Fit Objects}
\usage{
\method{plot}{vel_bmgm}(x, type = c("coefficients", "ppi", "graph", "trace"), ...)
}
\arguments{
\item{x}{A fit object from bmgm_GP().}

\item{type}{Character; plot type. One of "coefficients" (default), "ppi", "graph", "trace".}

\item{...}{Additional arguments passed to the underlying plot function.}
}
\value{
A ggplot object (or base plot for "graph").
}
\description{
Dispatches to the appropriate VEL-BMGM visualization based on the type argument.
}
